+++ stats.jl
@@ -44,18 +44,18 @@ end
 # documenting scores
 function initscoresdict(n_folds; include_losses=false)
     scores_dict = Dict{String, Any}(
-        "r2score_train"=>Vector{Float64}(undef, n_folds),
-        "r2score_val"=>Vector{Float64}(undef, n_folds),
-        "adj_r2score_train"=>Vector{Float64}(undef, n_folds),
-        "adj_r2score_val"=>Vector{Float64}(undef, n_folds),
-        "mse_train"=>Vector{Float64}(undef, n_folds),
-        "mse_val"=>Vector{Float64}(undef, n_folds),
-        "mae_train"=>Vector{Float64}(undef, n_folds),
-        "mae_val"=>Vector{Float64}(undef, n_folds),
+        "r2score_train"=>Vector{Float32}(undef, n_folds),
+        "r2score_val"=>Vector{Float32}(undef, n_folds),
+        "adj_r2score_train"=>Vector{Float32}(undef, n_folds),
+        "adj_r2score_val"=>Vector{Float32}(undef, n_folds),
+        "mse_train"=>Vector{Float32}(undef, n_folds),
+        "mse_val"=>Vector{Float32}(undef, n_folds),
+        "mae_train"=>Vector{Float32}(undef, n_folds),
+        "mae_val"=>Vector{Float32}(undef, n_folds),
     )
 
     if include_losses
-        scores_dict["training_losses"] = Vector{Vector{Float64}}(undef, n_folds)
+        scores_dict["training_losses"] = Vector{Vector{Float32}}(undef, n_folds)
     end
 
     return scores_dict
